Add files via upload
Welcome to Yashpreet's shipping calculator

1.	Create a Java class named Package that contains the following:
a.	Package should have three private instance variables of type double named length, width, and height. 
b.	Package should have one private instance variable of the type Scanner named input, initialized to System.in. 
c.	No-args (explicit default) public constructor, which initializes all three double instance variables to 1.0.  
d.	Initial (parameterized) public constructor, which defines three parameters of type double, named length, width, and height, which are used to initialize the instance variables of same name. 
e.	Public copy constructor, with a parameter of type Package, used to duplicate an existing Package object. 
f.	Three public void methods named inputLength, inputWidth, and inputHeight. Each method will prompt the user for the appropriate property, and input a double value using the Scanner object input to initialize the instance variables 
g.	A public void method named displayDimensions which prints the dimensions as length X width X height (each value separated by a “ X “). 
h.	A public method of type double named calcVolume that calculates the volume and returns the result as a double value. 

2.	Create a class named Shipment
a.	The program must ask for two separate packages to ship
b.	The program must calculate the cost difference using the difference in volume:
i.e.  The base price for a package with volume <=1 is $3, for every unit increase in volume, the cost increases by $1
e.g. 1: a parcel with volume 4, the cost is $3+ $1 +$1 +$1 = $6
e.g. 2: a parcel with volume 2.5, the cost is $3 + $1.5 = $4.5
c.	Give the following (in order of priority):  
i.	If there is no difference, display the costs as the same  
ii.	If the cost of one is less than twice the other, display that it is “slightly more than”  
iii.	If the cost of one is less than three times the other, display that it is “twice”  
iv.	If the cost of one is less than four times the other, display that it is “triple”  
v.	If the cost of one is less than five times the other, display that it is “quadruple”  
vi.	Otherwise, display that as a calculated multiple
c.	The program must indicate the more costly package (if not same cost) and by how much  
d.	The program must calculate and display the appropriate message (including proper dimension and cost format)  
f.	Your code also does NOT need to worry if the user inputs an invalid value for the input (example: invalid length). The output of your code must match the samples.